+++ srcs/asm/main_asm.c
@@ -19,14 +19,14 @@ int main(int ac, char **av)
 
 	fd = 0;
 	if (ac == 0)
-		return (ft_error("No arguments", NULL));
+		return ((int)ft_error("No arguments", NULL));
 	if (ac == 1)
 		return (usage(PROG_ASM));
 	if (ac > 2)
-		return (ft_error("Too many arguments", NULL));
+		return ((int)ft_error("Too many arguments", NULL));
 	fd = open(av[1], O_RDONLY);
 	if (fd < 0)
-		return (ft_error("A problem occured while opening the source file:", av[1]));
+		return ((int)ft_error("A problem occured while opening the source file:", av[1]));
 	if (cor_file(av[1], &out_file, fd) == FALSE)
 		return (FALSE);
 	write(out_file.fd, out_file.content, out_file.total_size);
